+++ abuse/trunk/src/imlib/include/jmalloc.hpp
@@ -7,14 +7,14 @@
 #ifdef MANAGE_MEM
 enum {ALLOC_SPACE_STATIC,ALLOC_SPACE_CACHE};
 extern int alloc_space;
-void *jmalloc(long size, char *what_for);
-void *jrealloc(void *ptr, long size, char *what_for);
+void *jmalloc(int32_t size, char *what_for);
+void *jrealloc(void *ptr, int32_t size, char *what_for);
 void jfree(void *ptr);
 void mem_report(char *filename);
-void jmalloc_init(long min_size);
+void jmalloc_init(int32_t min_size);
 void jmalloc_uninit();
-long j_allocated();
-long j_available();
+int32_t j_allocated();
+int32_t j_available();
 extern void free_up_memory();
 #else
 #define jmalloc(x,y) malloc(x)
